requirements.txt
flask
pillow
pytesseract
python-dotenv
click